+++ src/start-client.jsx
@@ -8,7 +8,6 @@ import { ReduxAsyncConnect } from 'redux-connect';
 import { loadableReady } from '@loadable/component';
 import routes from '~/routes';
 import { settings } from '~/config';
-import * as OfflinePluginRuntime from 'offline-plugin/runtime';
 import '~/theme';
 
 import configureStore from '@plone/volto/store';
@@ -21,11 +20,9 @@ export const history = createBrowserHistory();
 
 initSentry(Sentry);
 
-OfflinePluginRuntime.install({
-  onInstalled: () => {
-    console.log('Service worker installed');
-  },
-});
+if (process.env.NODE_ENV === 'production') {
+  import('offline-plugin/runtime').then((plugin) => plugin.install()); // eslint-disable-line global-require
+}
 
 function reactIntlErrorHandler(error) {
   if (process.env.NODE_ENV !== 'production') {